Move CompositeUnit class to inner class of Version2g.
